`Sha256Sum`も分離
