+++ crates/voicevox_core_c_api/tests/e2e/main.rs
@@ -8,6 +8,7 @@ use serde::{Deserialize, Serialize};
 use self::assert_cdylib::Utf8Output;
 
 mod assert_cdylib;
+mod hash;
 mod mask;
 mod snapshots;
 mod testcases;
